requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
jinja2>=3.1,<4.0
sqlalchemy>=2.0,<3.0
pydantic>=2.7,<3.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
akshare>=1.14,<2.0
python-multipart>=0.0.9,<1.0
PyYAML>=6.0,<7.0
tomli>=2.0,<3.0
httpx>=0.27,<1.0
pytest>=8.0,<9.0
